# perf: Make ViewModelBase's internal errors dictionary lazy initialized