GitHub Issue: #
<!-- Link to relevant GitHub issue if applicable.
     All PRs should be associated with an issue -->

## Proposed Changes
<!-- Please check one or more that apply to this PR. -->

 - [ ] Bug fix
 - [ ] Feature
 - [ ] Code style update (formatting)
 - [ ] Refactoring (no functional changes, no api changes)
 - [ ] Build or CI related changes
 - [ ] Documentation content changes
 - [x] Other, please describe: Performance optimization

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying,
     or link to a relevant issue. -->

- The internal `_errors` dictionary for the `INotifyDataErrorInfo` aspect of `ViewModelBase` is always instantiated, even if the `INotifyDataErrorInfo` features aren't used for a ViewModel.
- The `_errors` dictionary is a `ConcurrentDictionary`.

| Method                              | Mean      | Allocated |
|------------------------------------ |----------:|----------:|
| CreateViewModel                     |  2.057 μs |    1760 B |
| CreateViewModel_WithExplicitName    |  2.077 μs |    1760 B |

## What is the new behavior?
<!-- Please describe the new behavior after your modifications. -->

- The `_errors` field is now a `Dictionary`. We weren't using any features specific to `ConcurrentDictionary`.
- The `_errors` field is now lazy-initialized. It stays `null` as long a we don't call `SetErrors`.

| Method                              | Mean      | Allocated |
|------------------------------------ |----------:|----------:|
| CreateViewModel                     |  1.280 μs |     936 B |
| CreateViewModel_WithExplicitName    |  1.193 μs |     936 B |

## Impact on version
<!-- Please select one or more based on your commits. -->

- [ ] **Major** (Public API was modified.)
  - Public constructs (class, struct, delegate, enum, etc.) were removed or renamed.
  - Public members were removed or renamed.
  - Public method signatures were changed or renamed.
- [ ] **Minor** (Public API was extended.)
  - Public constructs, members, or overloads were added.
- [x] **Patch** (Public API was unchanged.)
  - A bug in behavior was fixed.
  - Documentation was changed.
- [ ] **None** (The library is unchanged.)
  - Only code under the `build` folder was changed.
  - Only code under the `.github` folder was changed.

## Checklist

Please check that your PR fulfills the following requirements:

- [ ] Documentation has been added/updated.
- [x] Automated Unit / Integration tests for the changes have been added/updated.
- [ ] Updated [BREAKING_CHANGES.md](../BREAKING_CHANGES.md) (if you introduced a breaking change).
- [x] Your conventional commits are aligned with the **Impact on version** section.

<!-- If this PR contains a breaking change, please describe the impact
     and migration path for existing applications below. -->

## Other information
<!-- Please provide any additional information if necessary -->